requirements.txt
# Django Backend Requirements
# Python 3.12+

# Django Framework
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0

# REST Framework & JWT Authentication
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# WebSocket Support (Django Channels)
channels==4.3.2
daphne==4.2.1

# Database
psycopg2-binary==2.9.11

# Async & Real-time Support
autobahn==25.12.2
twisted==25.5.0
txaio==25.12.2

# Security & Encryption
cryptography==46.0.3
pyOpenSSL==25.3.0
pyasn1==0.6.1
pyasn1-modules==0.4.2

# Other Dependencies
attrs==25.4.0
automat==25.4.16
cffi==2.0.0
constantly==23.10.4
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
msgpack==1.1.2
packaging==25.0
pycparser==2.23
service-identity==24.2.0
sqlparse==0.5.5
typing-extensions==4.15.0
ujson==5.11.0
zope.interface==8.1.1

# Serialization Formats
cbor2==5.7.1
py-ubjson==0.16.1

